+++ Ash2/src/System/PlayerMotion/Landing.cpp
@@ -1,0 +1,24 @@
+#include <Siv3D.hpp>
+
+#include "Phase/FrameData.hpp"
+#include "System/PlayerMotion/Helper.hpp"
+#include "System/PlayerMotionSystem.hpp"
+
+namespace PlayerMotion {
+
+Optional<Motion> Tick(Landing& state, entt::registry& registry,
+                      entt::entity entity, const FrameData& frameData) {
+  StopHorizontalMovement(registry, entity);
+
+  auto& anim = registry.get<SpriteAnimation>(entity);
+  SetClip(anim, U"landing");
+
+  state.timer -= frameData.dt;
+  if (state.timer <= 0.0) {
+    return Neutral{};
+  }
+
+  return none;
+}
+
+}  // namespace PlayerMotion
